requirements.txt
snscrape
requests
pyyaml
yfinance
transformers
torch
backtrader
technical-analysis-library
discord-webhook
apscheduler
streamlit
